+++ python/egglog/runtime.py
@@ -11,12 +11,13 @@ so they are not mangled by Python and can be accessed by the user.
 
 from __future__ import annotations
 
+import itertools
 import operator
 from collections.abc import Callable
 from dataclasses import dataclass, replace
 from inspect import Parameter, Signature
 from itertools import zip_longest
-from typing import TYPE_CHECKING, TypeVar, Union, cast, get_args, get_origin
+from typing import TYPE_CHECKING, Any, TypeVar, Union, cast, get_args, get_origin
 
 from .declarations import *
 from .pretty import *
@@ -27,15 +28,14 @@ from .version_compat import *
 if TYPE_CHECKING:
     from collections.abc import Iterable
 
-    from .egraph import Fact
-
 
 __all__ = [
     "LIT_CLASS_NAMES",
-    "REFLECTED_BINARY_METHODS",
+    "NUMERIC_BINARY_METHODS",
     "RuntimeClass",
     "RuntimeExpr",
     "RuntimeFunction",
+    "define_expr_method",
     "resolve_callable",
     "resolve_type_annotation",
     "resolve_type_annotation_mutate",
@@ -46,24 +46,30 @@ UNIT_CLASS_NAME = "Unit"
 UNARY_LIT_CLASS_NAMES = {"i64", "f64", "Bool", "String"}
 LIT_CLASS_NAMES = UNARY_LIT_CLASS_NAMES | {UNIT_CLASS_NAME, "PyObject"}
 
-REFLECTED_BINARY_METHODS = {
-    "__radd__": "__add__",
-    "__rsub__": "__sub__",
-    "__rmul__": "__mul__",
-    "__rmatmul__": "__matmul__",
-    "__rtruediv__": "__truediv__",
-    "__rfloordiv__": "__floordiv__",
-    "__rmod__": "__mod__",
-    "__rpow__": "__pow__",
-    "__rlshift__": "__lshift__",
-    "__rrshift__": "__rshift__",
-    "__rand__": "__and__",
-    "__rxor__": "__xor__",
-    "__ror__": "__or__",
+# All methods which should return NotImplemented if they fail to resolve and are reflected as well
+# From https://docs.python.org/3/reference/datamodel.html#emulating-numeric-types
+
+NUMERIC_BINARY_METHODS = {
+    "__add__",
+    "__sub__",
+    "__mul__",
+    "__matmul__",
+    "__truediv__",
+    "__floordiv__",
+    "__mod__",
+    "__divmod__",
+    "__pow__",
+    "__lshift__",
+    "__rshift__",
+    "__and__",
+    "__xor__",
+    "__or__",
 }
 
-# Methods that need to return real Python values not expressions
-PRESERVED_METHODS = [
+
+# Methods that need to be defined on the runtime type that holds `Expr` objects, so that they can be used as methods.
+
+TYPE_DEFINED_METHODS = {
     "__bool__",
     "__len__",
     "__complex__",
@@ -71,9 +77,18 @@ PRESERVED_METHODS = [
     "__float__",
     "__iter__",
     "__index__",
-    "__float__",
-    "__int__",
-]
+    "__call__",
+    "__getitem__",
+    "__setitem__",
+    "__delitem__",
+    "__pos__",
+    "__neg__",
+    "__invert__",
+    "__lt__",
+    "__le__",
+    "__gt__",
+    "__ge__",
+}
 
 # Set this globally so we can get access to PyObject when we have a type annotation of just object.
 # This is the only time a type annotation doesn't need to include the egglog type b/c object is top so that would be redundant statically.
@@ -288,6 +303,14 @@ class RuntimeClass(DelayedDeclerations):
     def __hash__(self) -> int:
         return hash((id(self.__egg_decls_thunk__), self.__egg_tp__))
 
+    def __eq__(self, other: object) -> bool:
+        """
+        Support equality for runtime comparison of egglog classes.
+        """
+        if not isinstance(other, RuntimeClass):
+            return NotImplemented
+        return self.__egg_tp__ == other.__egg_tp__
+
     # Support unioning like types
     def __or__(self, value: type) -> object:
         return Union[self, value]  # noqa: UP007
@@ -357,7 +380,7 @@ class RuntimeFunction(DelayedDeclerations):
         try:
             bound = py_signature.bind(*args, **kwargs)
         except TypeError as err:
-            raise TypeError(f"Failed to call {self} with args {args} and kwargs {kwargs}") from err
+            raise TypeError(f"Failed to bind arguments for {self} with args {args} and kwargs {kwargs}: {err}") from err
         del kwargs
         bound.apply_defaults()
         assert not bound.kwargs
@@ -437,32 +460,6 @@ def to_py_signature(sig: FunctionSignature, decls: Declarations, optional_args: 
     return Signature(parameters)
 
 
-# All methods which should return NotImplemented if they fail to resolve
-# From https://docs.python.org/3/reference/datamodel.html
-PARTIAL_METHODS = {
-    "__lt__",
-    "__le__",
-    "__eq__",
-    "__ne__",
-    "__gt__",
-    "__ge__",
-    "__add__",
-    "__sub__",
-    "__mul__",
-    "__matmul__",
-    "__truediv__",
-    "__floordiv__",
-    "__mod__",
-    "__divmod__",
-    "__pow__",
-    "__lshift__",
-    "__rshift__",
-    "__and__",
-    "__xor__",
-    "__or__",
-}
-
-
 @dataclass
 class RuntimeExpr(DelayedDeclerations):
     __egg_typed_expr_thunk__: Callable[[], TypedExprDecl]
@@ -479,17 +476,14 @@ class RuntimeExpr(DelayedDeclerations):
         return self.__egg_typed_expr_thunk__()
 
     def __getattr__(self, name: str) -> RuntimeFunction | RuntimeExpr | Callable | None:
-        cls_name = self.__egg_class_name__
-        class_decl = self.__egg_class_decl__
-
-        if name in (preserved_methods := class_decl.preserved_methods):
-            return preserved_methods[name].__get__(self)
-
-        if name in class_decl.methods:
-            return RuntimeFunction(self.__egg_decls_thunk__, Thunk.value(MethodRef(cls_name, name)), self)
-        if name in class_decl.properties:
-            return RuntimeFunction(self.__egg_decls_thunk__, Thunk.value(PropertyRef(cls_name, name)), self)()
-        raise AttributeError(f"{cls_name} has no method {name}") from None
+        if (method := _get_expr_method(self, name)) is not None:
+            return method
+        if name in self.__egg_class_decl__.properties:
+            fn = RuntimeFunction(
+                self.__egg_decls_thunk__, Thunk.value(PropertyRef(self.__egg_class_name__, name)), self
+            )
+            return fn()
+        raise AttributeError(f"{self.__egg_class_name__} has no method {name}") from None
 
     def __repr__(self) -> str:
         """
@@ -520,13 +514,6 @@ class RuntimeExpr(DelayedDeclerations):
     def __egg_class_decl__(self) -> ClassDecl:
         return self.__egg_decls__.get_class_decl(self.__egg_class_name__)
 
-    # These both will be overriden below in the special methods section, but add these here for type hinting purposes
-    def __eq__(self, other: object) -> Fact:  # type: ignore[override, empty-body]
-        ...
-
-    def __ne__(self, other: object) -> RuntimeExpr:  # type: ignore[override, empty-body]
-        ...
-
     # Implement these so that copy() works on this object
     # otherwise copy will try to call `__getstate__` before object is initialized with properties which will cause inifinite recursion
 
@@ -540,89 +527,127 @@ class RuntimeExpr(DelayedDeclerations):
     def __hash__(self) -> int:
         return hash(self.__egg_typed_expr__)
 
+    # Implement this directly to special case behavior where it transforms to an egraph equality, if it is not a
+    # preserved method or defined on the class
+    def __eq__(self, other: object) -> object:  # type: ignore[override]
+        if (method := _get_expr_method(self, "__eq__")) is not None:
+            return method(other)
 
-# Define each of the special methods, since we have already declared them for pretty printing
-for name in list(BINARY_METHODS) + list(UNARY_METHODS) + ["__getitem__", "__call__", "__setitem__", "__delitem__"]:
+        # TODO: Check if two objects can be upcasted to be the same. If not, then return NotImplemented so other
+        # expr gets a chance to resolve __eq__ which could be a preserved method.
+        from .egraph import BaseExpr, eq
 
-    def _special_method(
-        self: RuntimeExpr,
-        *args: object,
-        __name: str = name,
-        **kwargs: object,
-    ) -> RuntimeExpr | Fact | None:
-        from .conversion import ConvertError
+        return eq(cast("BaseExpr", self)).to(cast("BaseExpr", other))
 
-        class_name = self.__egg_class_name__
-        class_decl = self.__egg_class_decl__
-        # First, try to resolve as preserved method
-        try:
-            method = class_decl.preserved_methods[__name]
-        except KeyError:
-            pass
-        else:
-            return method(self, *args, **kwargs)
-        # If this is a "partial" method meaning that it can return NotImplemented,
-        # we want to find the "best" superparent (lowest cost) of the arg types to call with it, instead of just
-        # using the arg type of the self arg.
-        # This is neccesary so if we add like an int to a ndarray, it will upcast the int to an ndarray, instead of vice versa.
-        if __name in PARTIAL_METHODS:
-            try:
-                return call_method_min_conversion(self, args[0], __name)
-            except ConvertError:
-                # Defer raising not imeplemented in case the dunder method is not symmetrical, then
-                # we use the standard process
-                pass
-        if __name in class_decl.methods:
-            fn = RuntimeFunction(self.__egg_decls_thunk__, Thunk.value(MethodRef(class_name, __name)), self)
-            return fn(*args, **kwargs)  # type: ignore[arg-type]
-        # Handle == and != fallbacks to eq and ne helpers if the methods aren't defined on the class explicitly.
-        if __name == "__eq__":
-            from .egraph import BaseExpr, eq
+    def __ne__(self, other: object) -> object:  # type: ignore[override]
+        if (method := _get_expr_method(self, "__ne__")) is not None:
+            return method(other)
 
-            return eq(cast("BaseExpr", self)).to(cast("BaseExpr", args[0]))
-        if __name == "__ne__":
-            from .egraph import BaseExpr, ne
+        from .egraph import BaseExpr, ne
 
-            return cast("RuntimeExpr", ne(cast("BaseExpr", self)).to(cast("BaseExpr", args[0])))
+        return ne(cast("BaseExpr", self)).to(cast("BaseExpr", other))
 
-        if __name in PARTIAL_METHODS:
-            return NotImplemented
-        raise TypeError(f"{class_name!r} object does not support {__name}")
-
-    setattr(RuntimeExpr, name, _special_method)
-
-# For each of the reflected binary methods, translate to the corresponding non-reflected method
-for reflected, non_reflected in REFLECTED_BINARY_METHODS.items():
-
-    def _reflected_method(self: RuntimeExpr, other: object, __non_reflected: str = non_reflected) -> RuntimeExpr | None:
-        # All binary methods are also "partial" meaning we should try to upcast first.
-        return call_method_min_conversion(other, self, __non_reflected)
-
-    setattr(RuntimeExpr, reflected, _reflected_method)
+    def __call__(
+        self, *args: object, **kwargs: object
+    ) -> object:  # define it here only for type checking, it will be overriden below
+        ...
 
 
-def call_method_min_conversion(slf: object, other: object, name: str) -> RuntimeExpr | None:
-    from .conversion import min_convertable_tp, resolve_literal
+def _get_expr_method(expr: RuntimeExpr, name: str) -> RuntimeFunction | RuntimeExpr | Callable | None:
+    if name in (preserved_methods := expr.__egg_class_decl__.preserved_methods):
+        return preserved_methods[name].__get__(expr)
 
-    # find a minimum type that both can be converted to
-    # This is so so that calls like `-0.1 * Int("x")` work by upcasting both to floats.
-    min_tp = min_convertable_tp(slf, other, name).to_var()
-    slf = resolve_literal(min_tp, slf)
-    other = resolve_literal(min_tp, other)
-    method = RuntimeFunction(slf.__egg_decls_thunk__, Thunk.value(MethodRef(slf.__egg_class_name__, name)), slf)
-    return method(other)
+    if name in expr.__egg_class_decl__.methods:
+        return RuntimeFunction(expr.__egg_decls_thunk__, Thunk.value(MethodRef(expr.__egg_class_name__, name)), expr)
+    return None
 
 
-for name in PRESERVED_METHODS:
+def define_expr_method(name: str) -> None:
+    """
+    Given the name of a method, explicitly defines it on the runtime type that holds `Expr` objects as a method.
 
-    def _preserved_method(self: RuntimeExpr, __name: str = name):
-        try:
-            method = self.__egg_decls__.get_class_decl(self.__egg_typed_expr__.tp.name).preserved_methods[__name]
-        except KeyError as e:
-            raise TypeError(f"{self.__egg_typed_expr__.tp.name} has no method {__name}") from e
-        return method(self)
+    Call this if you need a method to be defined on the type itself where overrindg with `__getattr__` does not suffice,
+    like for NumPy's `__array_ufunc__`.
+    """
 
-    setattr(RuntimeExpr, name, _preserved_method)
+    def _defined_method(self: RuntimeExpr, *args, __name: str = name, **kwargs):
+        fn = _get_expr_method(self, __name)
+        if fn is None:
+            raise TypeError(f"{self.__egg_class_name__} expression has no method {__name}")
+        return fn(*args, **kwargs)
+
+    setattr(RuntimeExpr, name, _defined_method)
+
+
+for name in TYPE_DEFINED_METHODS:
+    define_expr_method(name)
+
+
+for name, reversed in itertools.product(NUMERIC_BINARY_METHODS, (False, True)):
+
+    def _numeric_binary_method(self: object, other: object, name: str = name, reversed: bool = reversed) -> object:
+        """
+        Implements numeric binary operations.
+
+        Tries to find the minimum cost conversion of either the LHS or the RHS, by finding all methods with either
+        the LHS or the RHS as exactly the right type and then upcasting the other to that type.
+        """
+        # 1. switch if reversed
+        if reversed:
+            self, other = other, self
+        # If the types don't exactly match to start, then we need to try converting one of them, by finding the cheapest conversion
+        if not (
+            isinstance(self, RuntimeExpr)
+            and isinstance(other, RuntimeExpr)
+            and (
+                self.__egg_decls__.check_binary_method_with_types(
+                    name, self.__egg_typed_expr__.tp, other.__egg_typed_expr__.tp
+                )
+            )
+        ):
+            from .conversion import CONVERSIONS, resolve_type, retrieve_conversion_decls
+
+            # tuple of (cost, convert_self)
+            best_method: (
+                tuple[
+                    int,
+                    Callable[[Any], RuntimeExpr],
+                ]
+                | None
+            ) = None
+            # Start by checking if we have a LHS that matches exactly and a RHS which can be converted
+            if (
+                isinstance(self, RuntimeExpr)
+                and (
+                    desired_other_type := self.__egg_decls__.check_binary_method_with_self_type(
+                        name, self.__egg_typed_expr__.tp
+                    )
+                )
+                and (converter := CONVERSIONS.get((resolve_type(other), desired_other_type)))
+            ):
+                best_method = (converter[0], lambda x: x)
+
+            # Next see if it's possible to convert the LHS and keep the RHS as is
+            if isinstance(other, RuntimeExpr):
+                decls = retrieve_conversion_decls()
+                other_type = other.__egg_typed_expr__.tp
+                resolved_self_type = resolve_type(self)
+                for desired_self_type in decls.check_binary_method_with_other_type(name, other_type):
+                    if converter := CONVERSIONS.get((resolved_self_type, desired_self_type)):
+                        cost, convert_self = converter
+                        if best_method is None or best_method[0] > cost:
+                            best_method = (cost, convert_self)
+
+            if not best_method:
+                raise RuntimeError(f"Cannot resolve {name} for {self} and {other}, no conversion found")
+            self = best_method[1](self)
+
+        method_ref = MethodRef(self.__egg_class_name__, name)
+        fn = RuntimeFunction(Thunk.value(self.__egg_decls__), Thunk.value(method_ref), self)
+        return fn(other)
+
+    method_name = f"__r{name[2:]}" if reversed else name
+    setattr(RuntimeExpr, method_name, _numeric_binary_method)
 
 
 def resolve_callable(callable: object) -> tuple[CallableRef, Declarations]: